refactor DiscordWebhook remove its class.
